+++ LOTECH Audio Metadata Editor/Models/AppDetailMode.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum AppDetailMode {
+    case singleFile
+    case batchEdit
+}